Create LinkCheckerMT.java
Initial version.
This program checks the http links found in a web page. Program takes one argument which is the path of the properties file. In the properties file, there must be a property named "URL" pointing to the web page to test the links on. When program is complete, if there are any "bad links", a list of the bad links will be produced.