Create Stringify.java
This class is a utility that uses Java reflection to print out the string contents, to the best extent possible, of the individual fields of an object. An example of usage is to print such data to a log file. In addition, there is the capability to either exclude the contents of certain fields from being emitted or only include the contents of certain fields to be emitted.